Split ReceiverController: extract peer-link sessions
Second in the receiver-split arc (#809 was the cleanup loop).
Extracts the peer-link session lifecycle plus the
queue-status broadcast cluster into a new sibling module
controllers/remote_build/peer_link_sessions.py:

- register_peer_link_session - registers a new session,
  evicts a stale same-dashboard_id slot via SUPERSEDED,
  pushes an initial queue_status push.
- unregister_peer_link_session - drops a session from
  the registry, fires CLOSED.
- handle_cancel_job - receiver-side dispatch for inbound
  cancel_job peer-link frames.
- on_firmware_queue_transition - bus listener that
  schedules a per-session queue_status broadcast on every
  firmware lifecycle event.
- broadcast_queue_status - fans the snapshot out to
  every active session.
- _send_initial_queue_status - single-session
  counterpart used by register_peer_link_session;
  module-private (only the register path uses it).

The four public entry points stay as bound-method delegates
on the controller (start() registers the firmware listener
with self._on_firmware_queue_transition, peer-link
session entry points are part of the cross-controller
surface). Bodies take ReceiverController as the first
arg; sibling state goes through controller.state.X.

Imports moved out of receiver.py: cast,
frame_schema / is_valid_frame, QueueStatusFrameData,
ReceiverPeerLinkSessionClosedData /
ReceiverPeerLinkSessionOpenedData, the
_CANCEL_JOB_SCHEMA constant.

Test repoint: 1 test that called
controller.receiver._broadcast_queue_status(...) directly
moved to rb_peer_link_sessions.broadcast_queue_status(receiver, ...).
The other 3 tests calling _on_firmware_queue_transition
go through the bound delegate unchanged.

Receiver: 1043 -> 842 lines (-201). Combined with #809:
1083 -> 842 (-241, -22%). All 3412 tests pass.
